Move formatter into own file.
